+++ binding/Binding/SKCanvas.cs
@@ -70,7 +70,7 @@ namespace SkiaSharp
 
 		public void DrawColor (SKColor color, SKBlendMode mode = SKBlendMode.Src)
 		{
-			SkiaApi.sk_canvas_draw_color (Handle, color, mode);
+			SkiaApi.sk_canvas_draw_color (Handle, (uint)color, mode);
 		}
 
 		public void DrawLine (SKPoint p0, SKPoint p1, SKPaint paint)
